+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ## Unreleased (4.4)
 
 - Fixed a JavaScript error that occurred when opening an asset selector modal, preventing the index results from showing initially. ([#12755](https://github.com/craftcms/cms/issues/12755))
+- Fixed a bug where it wasn’t possible to vertically scroll element index filter HUDs. ([#12751](https://github.com/craftcms/cms/pull/12751))
 
 ## 4.4.0-beta.3 - 2023-02-25